파이어 베이스 수정 (14)

+++ libmodule/src/main/java/com/dev/voltsoft/lib/firebase/db/FireBaseDBRequest.java
@@ -386,7 +386,7 @@ public class FireBaseDBRequest<T> extends BaseRequest implements Runnable
         }
         else
         {
-            EqualStartValue = (double) param2;
+            EqualStartValue = (int) param2;
         }
     }
